Move output templates into their own files
This makes working with the templates much nicer because it's no
longer a long python string.

Also it allows us to simplify the logic and makes adding new
implementations easier.
