+++ generate.py
@@ -42,419 +42,6 @@ MAX_CODEPOINT = 0x10FFFF
 
 CPP_PREFIX = "widechar_"
 
-OUTPUT_FILENAME = "widechar_width.h"
-OUTPUT_FILENAME_JS = "widechar_width.js"
-OUTPUT_FILENAME_RUST = "widechar_width.rs"
-
-OUTPUT_TEMPLATE = r"""
-/**
- * {filename}, generated on {today}.
- * See https://github.com/ridiculousfish/widecharwidth/
- *
- * SHA1 file hashes:
- *  UnicodeData.txt:     {unicode_hash}
- *  EastAsianWidth.txt:  {eaw_hash}
- *  emoji-data.txt:      {emoji_hash}
- */
-
-#ifndef WIDECHAR_WIDTH_H
-#define WIDECHAR_WIDTH_H
-
-#include <algorithm>
-#include <iterator>
-#include <cstddef>
-#include <cstdint>
-
-namespace {{
-
-/* Special width values */
-enum {{
-  {p}nonprint = -1,     // The character is not printable.
-  {p}combining = -2,    // The character is a zero-width combiner.
-  {p}ambiguous = -3,    // The character is East-Asian ambiguous width.
-  {p}private_use = -4,  // The character is for private use.
-  {p}unassigned = -5,   // The character is unassigned.
-  {p}widened_in_9 = -6, // Width is 1 in Unicode 8, 2 in Unicode 9+.
-  {p}non_character = -7 // The character is a noncharacter.
-}};
-
-/* An inclusive range of characters. */
-struct {p}range {{
-  uint32_t lo;
-  uint32_t hi;
-}};
-
-/* Simple ASCII characters - used a lot, so we check them first. */
-static const struct {p}range {p}ascii_table[] = {{
-    {ascii}
-}};
-
-/* Private usage range. */
-static const struct {p}range {p}private_table[] = {{
-    {private}
-}};
-
-/* Nonprinting characters. */
-static const struct {p}range {p}nonprint_table[] = {{
-    {nonprint}
-}};
-
-/* Width 0 combining marks. */
-static const struct {p}range {p}combining_table[] = {{
-    {combining}
-}};
-
-/* Width 0 combining letters. */
-static const struct {p}range {p}combiningletters_table[] = {{
-    {combiningletters}
-}};
-
-/* Width 2 characters. */
-static const struct {p}range {p}doublewide_table[] = {{
-    {doublewide}
-}};
-
-/* Ambiguous-width characters. */
-static const struct {p}range {p}ambiguous_table[] = {{
-    {ambiguous}
-}};
-
-/* Unassigned characters. */
-static const struct {p}range {p}unassigned_table[] = {{
-    {unassigned}
-}};
-
-/* Non-characters. */
-static const struct {p}range {p}nonchar_table[] = {{
-    {noncharacters}
-}};
-
-/* Characters that were widened from width 1 to 2 in Unicode 9. */
-static const struct {p}range {p}widened_table[] = {{
-    {widenedin9}
-}};
-
-template<typename Collection>
-bool {p}in_table(const Collection &arr, uint32_t c) {{
-    auto where = std::lower_bound(std::begin(arr), std::end(arr), c,
-        []({p}range p, uint32_t c) {{ return p.hi < c; }});
-    return where != std::end(arr) && where->lo <= c;
-}}
-
-/* Return the width of character c, or a special negative value. */
-int {p}wcwidth(uint32_t c) {{
-    if ({p}in_table({p}ascii_table, c))
-        return 1;
-    if ({p}in_table({p}private_table, c))
-        return {p}private_use;
-    if ({p}in_table({p}nonprint_table, c))
-        return {p}nonprint;
-    if ({p}in_table({p}nonchar_table, c))
-        return {p}non_character;
-    if ({p}in_table({p}combining_table, c))
-        return {p}combining;
-    if ({p}in_table({p}combiningletters_table, c))
-        return {p}combining;
-    if ({p}in_table({p}doublewide_table, c))
-        return 2;
-    if ({p}in_table({p}ambiguous_table, c))
-        return {p}ambiguous;
-    if ({p}in_table({p}unassigned_table, c))
-        return {p}unassigned;
-    if ({p}in_table({p}widened_table, c))
-        return {p}widened_in_9;
-    return 1;
-}}
-
-}} // namespace
-#endif // WIDECHAR_WIDTH_H
-"""
-
-OUTPUT_TEMPLATE_RUST = r"""
-/**
- * {filename}, generated on {today}.
- * See https://github.com/ridiculousfish/widecharwidth/
- *
- * SHA1 file hashes:
- *  UnicodeData.txt:     {unicode_hash}
- *  EastAsianWidth.txt:  {eaw_hash}
- *  emoji-data.txt:      {emoji_hash}
- */
-
-type R = (u32, u32);
-
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum WcWidth {{
-  /// The character is single-width
-  One,
-  /// The character is double-width
-  Two,
-  /// The character is not printable.
-  NonPrint,
-  /// The character is a zero-width combiner.
-  Combining,
-  /// The character is East-Asian ambiguous width.
-  Ambiguous,
-  /// The character is for private use.
-  PrivateUse,
-  /// The character is unassigned.
-  Unassigned,
-  /// Width is 1 in Unicode 8, 2 in Unicode 9+.
-  WidenedIn9,
-  /// The character is a noncharacter.
-  NonCharacter,
-}}
-
-/// Simple ASCII characters - used a lot, so we check them first.
-const ASCII_TABLE: &'static [R] = &[
-    {ascii}
-];
-
-/// Private usage range.
-const PRIVATE_TABLE: &'static [R] = &[
-    {private}
-];
-
-/// Nonprinting characters.
-const NONPRINT_TABLE: &'static [R] = &[
-    {nonprint}
-];
-
-/// Width 0 combining marks.
-const COMBINING_TABLE: &'static [R] = &[
-    {combining}
-];
-
-/// Width 0 combining letters.
-const COMBININGLETTERS_TABLE: &'static [R] = &[
-    {combiningletters}
-];
-
-/// Width 2 characters.
-const DOUBLEWIDE_TABLE: &'static [R] = &[
-    {doublewide}
-];
-
-/// Ambiguous-width characters.
-const AMBIGUOUS_TABLE: &'static [R] = &[
-    {ambiguous}
-];
-
-/// Unassigned characters.
-const UNASSIGNED_TABLE: &'static [R] = &[
-    {unassigned}
-];
-
-/// Non-characters.
-const NONCHAR_TABLE: &'static [R] = &[
-    {noncharacters}
-];
-
-/// Characters that were widened from width 1 to 2 in Unicode 9.
-const WIDENED_TABLE: &'static [R] = &[
-    {widenedin9}
-];
-
-fn in_table(arr: &[R], c: u32) -> bool {{
-    arr.binary_search_by(|(start, end)| if c >= *start && c <= *end {{
-        std::cmp::Ordering::Equal
-    }} else {{
-        start.cmp(&c)
-    }}).is_ok()
-}}
-
-impl WcWidth {{
-    /// Return the width of character c
-    pub fn from_char(c: char) -> Self {{
-        let c = c as u32;
-        if in_table(&ASCII_TABLE, c) {{
-            return Self::One;
-        }}
-        if in_table(&PRIVATE_TABLE, c) {{
-            return Self::PrivateUse;
-        }}
-        if in_table(&NONPRINT_TABLE, c) {{
-            return Self::NonPrint;
-        }}
-        if in_table(&NONCHAR_TABLE, c) {{
-            return Self::NonCharacter;
-        }}
-        if in_table(&COMBINING_TABLE, c) {{
-            return Self::Combining;
-        }}
-        if in_table(&COMBININGLETTERS_TABLE, c) {{
-            return Self::Combining;
-        }}
-        if in_table(&DOUBLEWIDE_TABLE, c) {{
-            return Self::Two;
-        }}
-        if in_table(&AMBIGUOUS_TABLE, c) {{
-            return Self::Ambiguous;
-        }}
-        if in_table(&UNASSIGNED_TABLE, c) {{
-            return Self::Unassigned;
-        }}
-        if in_table(&WIDENED_TABLE, c) {{
-            return Self::WidenedIn9;
-        }}
-        Self::One
-    }}
-
-    /// Returns width for applications that are using unicode 8 or earlier
-    pub fn width_unicode_8_or_earlier(self) -> u8 {{
-        match self {{
-            Self::One => 1,
-            Self::Two => 2,
-            Self::NonPrint | Self::Combining | Self::Unassigned | Self::NonCharacter => 0,
-            Self::Ambiguous | Self::PrivateUse => 1,
-            Self::WidenedIn9 => 1,
-        }}
-    }}
-
-    /// Returns width for applications that are using unicode 9 or later
-    pub fn width_unicode_9_or_later(self) -> u8 {{
-        if self == Self::WidenedIn9 {{
-            return 2;
-        }}
-        self.width_unicode_8_or_earlier()
-    }}
-}}
-
-#[cfg(test)]
-mod test {{
-    use super::*;
-
-    #[test]
-    fn basics() {{
-        assert_eq!(WcWidth::from_char('w'), WcWidth::One);
-        assert_eq!(WcWidth::from_char('\x1f'), WcWidth::NonPrint);
-        assert_eq!(WcWidth::from_char('\u{{e001}}'), WcWidth::PrivateUse);
-        assert_eq!(WcWidth::from_char('\u{{2716}}'), WcWidth::One);
-        assert_eq!(WcWidth::from_char('\u{{270a}}'), WcWidth::WidenedIn9);
-        assert_eq!(WcWidth::from_char('\u{{3fffd}}'), WcWidth::Two);
-    }}
-}}
-
-"""
-
-OUTPUT_TEMPLATE_JS = r"""
-/*
- * {filename}, generated on {today}.
- * See https://github.com/ridiculousfish/widecharwidth/
- *
- * SHA1 file hashes:
- *  UnicodeData.txt:     {unicode_hash}
- *  EastAsianWidth.txt:  {eaw_hash}
- *  emoji-data.txt:      {emoji_hash}
- */
-
-/* Special width values */
-const {p}nonprint = -1;     // The character is not printable.
-const {p}combining = -2;    // The character is a zero-width combiner.
-const {p}ambiguous = -3;    // The character is East-Asian ambiguous width.
-const {p}private_use = -4;  // The character is for private use.
-const {p}unassigned = -5;   // The character is unassigned.
-const {p}widened_in_9 = -6; // Width is 1 in Unicode 8, 2 in Unicode 9+.
-const {p}non_character = -7; // The character is a noncharacter.
-
-/* Simple ASCII characters - used a lot, so we check them first. */
-const {p}ascii_table = [
-    {ascii}
-];
-
-/* Private usage range. */
-const {p}private_table = [
-    {private}
-];
-
-/* Nonprinting characters. */
-const {p}nonprint_table = [
-    {nonprint}
-];
-
-/* Width 0 combining marks. */
-const {p}combining_table = [
-    {combining}
-];
-
-/* Width 0 combining letters. */
-const {p}combiningletters_table = [
-    {combiningletters}
-];
-
-/* Width.2 characters. */
-const {p}doublewide_table = [
-    {doublewide}
-];
-
-/* Ambiguous-width characters. */
-const {p}ambiguous_table = [
-    {ambiguous}
-];
-
-/* Unassigned characters. */
-const {p}unassigned_table = [
-    {unassigned}
-];
-
-/* Non-characters. */
-const {p}nonchar_table[] = [
-    {noncharacters}
-];
-
-/* Characters that were widened from width 1 to 2 in Unicode 9. */
-const {p}widened_table[] = [
-    {widenedin9}
-];
-
-function {p}in_table(data, ucs) {{
-    let min = 0;
-    let max = data.length - 1;
-    let mid;
-    if (ucs < data[0][0] || ucs > data[max][1])
-        return false;
-
-    while (max >= min) {{
-        mid = (min + max) >> 1;
-        if (ucs > data[mid][1]) {{
-            min = mid + 1;
-        }}
-        else if (ucs < data[mid][0]) {{
-            max = mid - 1;
-        }}
-        else {{
-            return true;
-        }}
-    }}
-    return false;
-}}
-
-/* Return the width of character c, or a special negative value. */
-function {p}wcwidth(c) {{
-    if ({p}in_table({p}ascii_table, c))
-        return 1;
-    if ({p}in_table({p}private_table, c))
-        return {p}private_use;
-    if ({p}in_table({p}nonprint_table, c))
-        return {p}nonprint;
-    if ({p}in_table({p}nonchar_table, c))
-        return {p}non_character;
-    if ({p}in_table({p}combining_table, c))
-        return {p}combining;
-    if ({p}in_table({p}combiningletters_table, c))
-        return {p}combining;
-    if ({p}in_table({p}doublewide_table, c))
-        return 2;
-    if ({p}in_table({p}ambiguous_table, c))
-        return {p}ambiguous;
-    if ({p}in_table({p}unassigned_table, c))
-        return {p}unassigned;
-    if ({p}in_table({p}widened_table, c))
-        return {p}widened_in_9;
-    return 1;
-}}
-"""
-
 # Ambiguous East Asian characters
 WIDTH_AMBIGUOUS_EASTASIAN = -3
 
@@ -775,20 +362,17 @@ def make_fields(
 if __name__ == "__main__":
     datas = read_datas()
     cps = make_codepoints(datas)
-    fields = make_fields(datas, cps, LangSettings("{}"), OUTPUT_FILENAME)
-    with open(OUTPUT_FILENAME, "w") as fd:
-        fd.write(OUTPUT_TEMPLATE.strip().format(**fields))
-        fd.write("\n")
-    log("Output " + OUTPUT_FILENAME)
-
-    fields = make_fields(datas, cps, LangSettings("[]"), OUTPUT_FILENAME_JS)
-    with open(OUTPUT_FILENAME_JS, "w") as fd:
-        fd.write(OUTPUT_TEMPLATE_JS.strip().format(**fields))
-        fd.write("\n")
-    log("Output " + OUTPUT_FILENAME_JS)
-
-    fields = make_fields(datas, cps, LangSettings("()"), OUTPUT_FILENAME_RUST)
-    with open(OUTPUT_FILENAME_RUST, "w") as fd:
-        fd.write(OUTPUT_TEMPLATE_RUST.strip().format(**fields))
-        fd.write("\n")
-    log("Output " + OUTPUT_FILENAME_RUST)
+    langs = {
+        ".h": "{}",
+        ".js": "[]",
+        ".rs": "()",
+    }
+    for suffix, settings in langs.items():
+        with open("templates/template" + suffix) as templatefile:
+            template = templatefile.read()
+            output = "widechar_width" + suffix
+            fields = make_fields(datas, cps, LangSettings(settings), output)
+            with open(output, "w") as fd:
+                fd.write(template.strip().format(**fields))
+                fd.write("\n")
+                log("Output " + output)
